chore: pass all tests; chore: remove long as pk;

+++ fcomb-frontend/src/main/scala/io/fcomb/frontend/api/Resource.scala
@@ -23,7 +23,7 @@ object Resource {
 
   val repositories = prefix + "/repositories"
   def repository(slug: String) = s"$repositories/$slug"
-  def repository(id: Long)     = s"$repositories/$id"
+  def repository(id: Int)      = s"$repositories/$id"
 
   val users  = prefix + "/users"
   val signUp = users + "/sign_up"
