+++ fcomb-models/jvm/src/main/scala/io/fcomb/models/docker/distribution/ImageBlob.scala
@@ -34,7 +34,7 @@ object ImageBlobState extends Enum[ImageBlobState] {
 
 case class ImageBlob(
     id: Option[UUID] = None,
-    imageId: Long,
+    imageId: Int,
     state: ImageBlobState,
     sha256Digest: Option[String],
     contentType: String,
