+++ fcomb-models/jvm/src/main/scala/io/fcomb/models/docker/distribution/ImageManifestLayer.scala
@@ -19,6 +19,6 @@ package io.fcomb.models.docker.distribution
 import java.util.UUID
 
 final case class ImageManifestLayer(
-    imageManifestId: Long,
+    imageManifestId: Int,
     layerBlobId: UUID
 )
